+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Timed stpp subtitles EBU-TT-D linePadding
+- Update dependencies
 
 ## [1.2.2] - 2024-03-05