+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Nothing yet
+
+## [1.3.0] - 2024-04-23
+
 ### Added
 
 - MPD Patch functionality with new `/patch_ttl` URL configuration
@@ -205,7 +209,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - features and URLs listed at livesim2 root page
 - configurable generated stpp subtitles with timing info
 
-[Unreleased]: https://github.com/Dash-Industry-Forum/livesim2/compare/v1.2.1...HEAD
+[Unreleased]: https://github.com/Dash-Industry-Forum/livesim2/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/Dash-Industry-Forum/livesim2/compare/v1.2.2...v1.3.0
+[1.2.2]: https://github.com/Dash-Industry-Forum/livesim2/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/Dash-Industry-Forum/livesim2/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/Dash-Industry-Forum/livesim2/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/Dash-Industry-Forum/livesim2/compare/v1.1.0...v1.1.1